Split image functions across library interface
